[Course statistics] Move components under pages
