+++ services/frontend/src/components/Routes/index.tsx
@@ -3,7 +3,6 @@ import { Navigate, Route, Routes as RouterRoutes } from 'react-router'
 
 import { isDefaultServiceProvider } from '@/common'
 import { CoursePopulation } from '@/components/CoursePopulation'
-import { CourseStatistics } from '@/components/CourseStatistics'
 import { CustomOpenUniPopulation } from '@/components/CustomOpenUniPopulation'
 import { CustomPopulation } from '@/components/CustomPopulation'
 import { LanguageCenterView } from '@/components/LanguageCenterView'
@@ -17,6 +16,7 @@ import { languageCenterViewEnabled } from '@/conf'
 import { Changelog } from '@/pages/Changelog'
 import { CloseToGraduation } from '@/pages/CloseToGraduation'
 import { CompletedCourses } from '@/pages/CompletedCoursesSearch'
+import { CourseStatistics } from '@/pages/CourseStatistics'
 import { Faculties } from '@/pages/Faculties'
 import { Feedback } from '@/pages/Feedback'
 import { FrontPage } from '@/pages/FrontPage'
