+++ services/frontend/src/pages/CourseStatistics/AttemptsTable/foldableRow.css
