+++ services/frontend/src/pages/CourseStatistics/CourseTable/courseTable.css
