+++ services/frontend/src/pages/CourseStatistics/ResultTabs/panes/Pane.jsx
