+++ services/frontend/src/pages/CourseStatistics/ResultTabs/panes/settings/ChartSettings.jsx
