+++ services/frontend/src/pages/CourseStatistics/ResultTabs/panes/settings/TableSettings.jsx
