+++ services/frontend/src/pages/CourseStatistics/ResultTabs/panes/settings/common/ProviderOrganization.jsx
