+++ services/frontend/src/pages/CourseStatistics/ResultTabs/panes/settings/common/Setting.jsx
