+++ services/frontend/src/pages/CourseStatistics/ResultTabs/panes/util.ts
