+++ services/frontend/src/pages/CourseStatistics/SingleCourseStats/countTotalStats.ts
