+++ services/frontend/src/pages/CourseStatistics/SingleCourseStats/index.jsx
@@ -5,13 +5,13 @@ import { useDispatch, useSelector } from 'react-redux'
 import { useLocation, useNavigate } from 'react-router'
 import { Button, Form, Grid, Header, Popup, Segment } from 'semantic-ui-react'
 
-import { ProgrammeDropdown } from '@/components/CourseStatistics/ProgrammeDropdown'
-import { ResultTabs } from '@/components/CourseStatistics/ResultTabs'
 import { useLanguage } from '@/components/LanguagePicker/useLanguage'
 import { useGetMaxYearsToCreatePopulationFromQuery } from '@/redux/populations'
 import { useGetSemestersQuery } from '@/redux/semesters'
 import { setSelectedCourse, clearSelectedCourse } from '@/redux/singleCourseStats'
 import { ALL, getAllStudyProgrammes } from '@/selectors/courseStats'
+import { ProgrammeDropdown } from '../ProgrammeDropdown'
+import { ResultTabs } from '../ResultTabs'
 import { countTotalStats } from './countTotalStats'
 import { YearFilter } from './YearFilter'
