+++ services/frontend/src/pages/CourseStatistics/SingleCourseTab/index.jsx
@@ -1,10 +1,10 @@
 import { useDispatch, useSelector } from 'react-redux'
 import { Divider, Form, Header, Label, Segment } from 'semantic-ui-react'
 
-import { SingleCourseStats } from '@/components/CourseStatistics/SingleCourseStats'
 import { useLanguage } from '@/components/LanguagePicker/useLanguage'
 import { setSelectedCourse } from '@/redux/singleCourseStats'
 import { getAvailableStats, getCourses, getCourseStats } from '@/selectors/courseStats'
+import { SingleCourseStats } from '../SingleCourseStats'
 
 const CourseSelector = ({ courses, selected, setSelected }) => {
   const dispatch = useDispatch()
