+++ services/frontend/src/pages/CourseStatistics/courseStatisticsUtils.js
