+++ services/frontend/src/pages/CourseStatistics/index.jsx
